Day 1
Doing some projects for Free Code Camp. projects need to be submitted through codepen so i'll add the link to the README later on. I sharpened my position skills in this project and used grids. I'm understanding the fundamentals now and breezing through, only had trouble with the nav bar toggle because idk javascript yet but was able to figure out a css alternative